refactor(session/stratum/server): session takes a message handler
Session now takes a message handler instead of using handleStratumMessage. This is necessary because
every session will use a different message handler, insofar as the protocol has a state which must
be tracked for each session. handleStratumMessage is now a higher-order function that takes request
handlers and returns a message handler.

Additionally, startSession has been removed and its functionality moved into the constructor of
Session. This is necessary because the session needs to know to remove itself from the sessions
object when it is destroyed, and it was originally put in there in startSession. If Session has to
clean up resources, it should also reserve those resources.