feat(server_session): reorganize protocol to handle general messages.
The request/response design that was used previously is insufficient
for general message handling in stratum. For example, some requests
need to be responded to with an error message followed by disconnection.
As another example, the subscribe method is followed by a response
and some notifications. We give the reuest handlers a more general
ability to interact with the protocol so as to respond in whatever
way they would like to requests. We also provide the ability to handle
notifications and responses, which aren't a big part of the server
session but will be important in the client session.