feat(subscribe): allow zero parameters for subscribe request
According to my reading of the Stratum protocol specs, the subscribe
request is supposed to have one required parameter (user-agent) and
one optional (extra-nonce-1). However, the poolin cpuminer software
sends a subscribe request with no parameters. Thus, we have made the
user-agent string optional. This makes sense to me because why do we
need to know that?